Split TeachingStaffProfile into 2 files
- Add TeachingStaffProfileEdit file
- Change APIRequester class into ApiRequester
