+++ React/central-student-hub/src/UserProfileComponent/Models/ContactInfo.ts
@@ -1,4 +1,4 @@
-type ContactInfo = {
+export type ContactInfo = {
     key: string,
     value: string
 }
