+++ React/central-student-hub/src/UserProfileComponent/Models/TeachingStaffProfileInfo.ts
@@ -1,4 +1,5 @@
 import { OfficeHours } from "./OfficeHours"
+import { ContactInfo } from "./ContactInfo"
 
 export type TeachingStaffProfileInfo = {
     teacherId: number,
