+++ React/central-student-hub/src/services/APIRequester.ts
@@ -1,8 +1,9 @@
-import { SignupRequest, SignupResponse } from './../UserSessionComponent/Models/SignupModels';
+import { SignupRequest, SignupResponse } from '../UserSessionComponent/Models/SignupModels.ts';
 import { LoginRequest, LoginResponse } from '../UserSessionComponent/Models/LoginModels.ts'
 import { FormState } from '../UserSessionComponent/Signup/Signup.tsx';
+import { TeachingStaffProfileInfo } from '../UserProfileComponent/Models/TeachingStaffProfileInfo.ts';
 
-export class APIRequester {
+export class ApiRequester {
     async login(request: LoginRequest): Promise<boolean> {
         try {
             const headers: HeadersInit = { "Content-Type": "application/json" };
@@ -36,11 +37,17 @@ export class APIRequester {
 
     async home(): Promise<string> {
         const token = document.cookie.split("=")[1];
-        console.log(token);
         const headers: HeadersInit = { "Authorization": `Bearer ${token}` };
         const requestOptions: RequestInit = { mode: 'cors', headers: headers, method: "get", credentials: "include" };
-        console.log(requestOptions);
         const response: Response = await fetch("http://localhost:8082/Hello", requestOptions);
         return await response.text();
+    }
+
+    async getTeachingStaffProfile(): Promise<TeachingStaffProfileInfo> {
+        const token = document.cookie.split("=")[1];
+        const headers: HeadersInit = { "Authorization": `Bearer ${token}` };
+        const requestOptions: RequestInit = { mode: 'cors', headers: headers, method: "get", credentials: "include" };
+        const response: Response = await fetch("http://localhost:8082/profile", requestOptions); // TODO: Change this to the correct endpoint
+        return await response.json();
     }
 }
